+++ .env
@@ -1,0 +1,3 @@
+FACEBOOK_APP_ID = 3038644956209663
+FACEBOOK_APP_SECRET = c2c25d92d0c992a9206c2ea088de85c1
+FACEBOOK_CALLBACK_URL = http://localhost:3000/users/auth/facebook/callback
