+++ app/controllers/users/omniauth_callbacks_controller.rb
@@ -1,0 +1,17 @@
+class Users::OmniauthCallbacksController < Devise::OmniauthCallbacksController
+    def facebook
+      @user = User.from_omniauth(request.env["omniauth.auth"])
+  
+      if @user.persisted?
+        sign_in_and_redirect @user, :event => :authentication
+        set_flash_message(:notice, :success, :kind => "Facebook") if is_navigational_format?
+      else
+        session["devise.facebook_data"] = request.env["omniauth.auth"]
+        redirect_to new_user_registration_url
+      end
+    end
+  
+    def failure
+      redirect_to root_path
+    end
+end
