+++ app/models/user.rb
@@ -21,7 +21,7 @@ class User < ApplicationRecord
 
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :validatable,
-         :omniauthable, omniauth_providers: %i[facebook]
+         :omniauthable, omniauth_providers: [:facebook]
   validates :email, presence: true
   validates :password, presence: true
   has_many :images
@@ -62,12 +62,19 @@ class User < ApplicationRecord
     confirmed_friends_posts + conirmed_inverse_friends_posts + posts
   end
 
+  def self.new_with_session(params, session)
+    super.tap do |user|
+      if data = session["devise.facebook_data"] && session["devise.facebook_data"]["extra"]["raw_info"]
+        user.email = data["email"] if user.email.blank?
+      end
+    end
+  end
+
   def self.from_omniauth(auth)
     where(provider: auth.provider, uid: auth.uid).first_or_create do |user|
       user.email = auth.info.email
-      user.password = devise.friendly_token[0, 20]
-      user.name = auth.info.name
-      user.image = auth.info.image
+      user.password = Devise.friendly_token[0,20]
+      user.username = auth.info.username   
     end
   end
 end
